bot_trader/pyproject.toml
[tool.poetry]
name = "bot-trader"
version = "0.1.0"
description = ""
authors = ["Mikhail Moskovkin <anstane228@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asyncpg = "^0.29.0"
aiogram = "^3.3.0"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
